Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/html/browsers/windows/browsing-context-names/browsing-context-choose-parent-004.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>HTML Test: Choose browsing context - '_parent' (case-sensitivity)</title>
<script src="../../../../../../../resources/testharness.js"></script>
<script src="../../../../../../../resources/testharnessreport.js"></script>
<div id="log"></div>
<iframe src="resources/parent-iframe-insensitive-1.html" name="parentWin" style="display:none"></iframe>

<script>
async_test(t => {
  window.addEventListener('message', t.step_func_done(e => {
    assert_equals(e.data.name, 'parentWin', "The browsing context name should be 'parentWin'.");
    assert_equals(e.data.isTop, false);
  }));
}, 'choosing _parent context should be case-insensitive');
</script>
